backend/requirements.txt
asgiref==3.10.0
colorama==0.4.6
Django==5.2.7
django-environ==0.12.0
django-role-permissions==3.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==37.12.0
gunicorn==23.0.0
iniconfig==2.3.0
isort==7.0.0
packaging==25.0
pluggy==1.6.0
psycopg==3.2.12
psycopg-binary==3.2.12
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
ruff==0.14.2
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
gunicorn>=22.0
django-extensions==4.1
django-rename-app==0.1.7
Pillow